Move networking api types to the api/types/networking package.
